requirements.txt
langchain
langchain-community
chromadb
pypdf
tiktoken
fastapi
uvicorn
requests
python-multipart
openai
langchain-pymupdf4llm
langchain-text-splitters
tqdm
sentence-transformers
chromadb
langchain-openrouter
python-dotenv
streamlit
uuid